+++ CHANGES.md
@@ -1,4 +1,7 @@
 # Changes
+### Unreleased
+- An hotfix was applied to ensure the support for Android 11 [RNMT-4515](https://outsystemsrd.atlassian.net/browse/RNMT-4515)
+
 ### cordova-sqlite-storage 3.2.0-OS1
 - Added support for async calls